Moving HSL and Pardiso to extensions
